Implemented sticky axes.
You can now make the axes 'stick' to the sides of the graphing panel
when they would otherwise appear off-screen. This is useful for when you
have something like a data set, but the points are so far off-screen
that the axes aren't visible.
You can enable it by right clicking the axes mathematical object and
selecting 'Sticky Axes'.